+++ mupdf/include/mupdf/fitz/path.h
@@ -70,7 +70,7 @@ void fz_trim_path(fz_context *ctx, fz_path *path);
 
 int fz_packed_path_size(const fz_path *path);
 
-int fz_pack_path(fz_context *ctx, uint8_t *pack, size_t max, const fz_path *path);
+size_t fz_pack_path(fz_context *ctx, uint8_t *pack, size_t max, const fz_path *path);
 
 fz_path *fz_clone_path(fz_context *ctx, fz_path *path);
