+++ mupdf/include/mupdf/fitz/writer.h
@@ -75,7 +75,7 @@ int fz_option_eq(const char *a, const char *b);
 /* Copy an option (val) into a destination buffer (dest), of maxlen bytes.
  * Returns the number of bytes (including terminator) that did not fit. If
  * val is maxlen or greater bytes in size, it will be left unterminated. */
-int fz_copy_option(fz_context *ctx, const char *val, char *dest, size_t maxlen);
+size_t fz_copy_option(fz_context *ctx, const char *val, char *dest, size_t maxlen);
 
 fz_document_writer *fz_new_document_writer(fz_context *ctx, const char *path, const char *format, const char *options);
