+++ mupdf/source/fitz/image.c
@@ -93,7 +93,7 @@ fz_cmp_image_key(fz_context *ctx, void *k0_, void *k1_)
 }
 
 static void
-fz_format_image_key(fz_context *ctx, char *s, int n, void *key_)
+fz_format_image_key(fz_context *ctx, char *s, size_t n, void *key_)
 {
 	fz_image_key *key = (fz_image_key *)key_;
 	fz_snprintf(s, n, "(image %d x %d sf=%d)", key->image->w, key->image->h, key->l2factor);
