+++ mupdf/source/pdf/pdf-op-run.c
@@ -909,7 +909,7 @@ pdf_show_space(fz_context *ctx, pdf_run_processor *pr, float tadj)
 }
 
 static void
-show_string(fz_context *ctx, pdf_run_processor *pr, unsigned char *buf, int len)
+show_string(fz_context *ctx, pdf_run_processor *pr, unsigned char *buf, size_t len)
 {
 	pdf_gstate *gstate = pr->gstate + pr->gtop;
 	pdf_font_desc *fontdesc = gstate->text.font;
@@ -933,7 +933,7 @@ show_string(fz_context *ctx, pdf_run_processor *pr, unsigned char *buf, int len)
 }
 
 static void
-pdf_show_string(fz_context *ctx, pdf_run_processor *pr, unsigned char *buf, int len)
+pdf_show_string(fz_context *ctx, pdf_run_processor *pr, unsigned char *buf, size_t len)
 {
 	pdf_gstate *gstate = pr->gstate + pr->gtop;
 	pdf_font_desc *fontdesc = gstate->text.font;
@@ -1714,13 +1714,13 @@ static void pdf_run_TJ(fz_context *ctx, pdf_processor *proc, pdf_obj *obj)
 	pdf_show_text(ctx, pr, obj);
 }
 
-static void pdf_run_Tj(fz_context *ctx, pdf_processor *proc, char *string, int string_len)
+static void pdf_run_Tj(fz_context *ctx, pdf_processor *proc, char *string, size_t string_len)
 {
 	pdf_run_processor *pr = (pdf_run_processor *)proc;
 	pdf_show_string(ctx, pr, (unsigned char *)string, string_len);
 }
 
-static void pdf_run_squote(fz_context *ctx, pdf_processor *proc, char *string, int string_len)
+static void pdf_run_squote(fz_context *ctx, pdf_processor *proc, char *string, size_t string_len)
 {
 	pdf_run_processor *pr = (pdf_run_processor *)proc;
 	pdf_gstate *gstate = pr->gstate + pr->gtop;
@@ -1728,7 +1728,7 @@ static void pdf_run_squote(fz_context *ctx, pdf_processor *proc, char *string, i
 	pdf_show_string(ctx, pr, (unsigned char*)string, string_len);
 }
 
-static void pdf_run_dquote(fz_context *ctx, pdf_processor *proc, float aw, float ac, char *string, int string_len)
+static void pdf_run_dquote(fz_context *ctx, pdf_processor *proc, float aw, float ac, char *string, size_t string_len)
 {
 	pdf_run_processor *pr = (pdf_run_processor *)proc;
 	pdf_gstate *gstate = pr->gstate + pr->gtop;
